+++ src/commands/run.rs
@@ -1,9 +1,11 @@
 //! Implementation of command line option for running server
 
-use crate::influxdb_ioxd::{self, serving_readiness::ServingReadinessState};
-use clap::arg_enum;
+use crate::{
+    influxdb_ioxd::{self, serving_readiness::ServingReadinessState},
+    object_store::ObjectStoreConfig,
+};
 use data_types::server_id::ServerId;
-use std::{net::SocketAddr, net::ToSocketAddrs, path::PathBuf};
+use std::{net::SocketAddr, net::ToSocketAddrs};
 use structopt::StructOpt;
 use thiserror::Error;
 use trogging::cli::{LoggingConfig, TracingConfig};
@@ -13,10 +15,6 @@ pub const DEFAULT_API_BIND_ADDR: &str = "127.0.0.1:8080";
 
 /// The default bind address for the gRPC.
 pub const DEFAULT_GRPC_BIND_ADDR: &str = "127.0.0.1:8082";
-
-/// The AWS region to use for Amazon S3 based object storage if none is
-/// specified.
-pub const FALLBACK_AWS_REGION: &str = "us-east-1";
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -80,6 +78,10 @@ pub struct Config {
     #[structopt(flatten)]
     pub(crate) tracing_config: TracingConfig,
 
+    // object store config
+    #[structopt(flatten)]
+    pub(crate) object_store_config: ObjectStoreConfig,
+
     /// The identifier for the server.
     ///
     /// Used for writing to object storage and as an identifier that is added to
@@ -107,10 +109,6 @@ pub struct Config {
     )]
     pub grpc_bind_address: SocketAddr,
 
-    /// The location InfluxDB IOx will use to store files locally.
-    #[structopt(long = "--data-dir", env = "INFLUXDB_IOX_DB_DIR")]
-    pub database_directory: Option<PathBuf>,
-
     /// The number of threads to use for all worker pools.
     ///
     /// IOx uses a pool with `--num-threads` threads *each* for
@@ -121,128 +119,6 @@ pub struct Config {
     /// If not specified, defaults to the number of cores on the system
     #[structopt(long = "--num-worker-threads", env = "INFLUXDB_IOX_NUM_WORKER_THREADS")]
     pub num_worker_threads: Option<usize>,
-
-    #[structopt(
-    long = "--object-store",
-    env = "INFLUXDB_IOX_OBJECT_STORE",
-    possible_values = &ObjectStore::variants(),
-    case_insensitive = true,
-    long_help = r#"Which object storage to use. If not specified, defaults to memory.
-
-Possible values (case insensitive):
-
-* memory (default): Effectively no object persistence.
-* memorythrottled: Like `memory` but with latency and throughput that somewhat resamble a cloud
-   object store. Useful for testing and benchmarking.
-* file: Stores objects in the local filesystem. Must also set `--data-dir`.
-* s3: Amazon S3. Must also set `--bucket`, `--aws-access-key-id`, `--aws-secret-access-key`, and
-   possibly `--aws-default-region`.
-* google: Google Cloud Storage. Must also set `--bucket` and `--google-service-account`.
-* azure: Microsoft Azure blob storage. Must also set `--bucket`, `--azure-storage-account`,
-   and `--azure-storage-access-key`.
-        "#,
-    )]
-    pub object_store: Option<ObjectStore>,
-
-    /// Name of the bucket to use for the object store. Must also set
-    /// `--object-store` to a cloud object storage to have any effect.
-    ///
-    /// If using Google Cloud Storage for the object store, this item as well
-    /// as `--google-service-account` must be set.
-    ///
-    /// If using S3 for the object store, must set this item as well
-    /// as `--aws-access-key-id` and `--aws-secret-access-key`. Can also set
-    /// `--aws-default-region` if not using the fallback region.
-    ///
-    /// If using Azure for the object store, set this item to the name of a
-    /// container you've created in the associated storage account, under
-    /// Blob Service > Containers. Must also set `--azure-storage-account` and
-    /// `--azure-storage-access-key`.
-    #[structopt(long = "--bucket", env = "INFLUXDB_IOX_BUCKET")]
-    pub bucket: Option<String>,
-
-    /// When using Amazon S3 as the object store, set this to an access key that
-    /// has permission to read from and write to the specified S3 bucket.
-    ///
-    /// Must also set `--object-store=s3`, `--bucket`, and
-    /// `--aws-secret-access-key`. Can also set `--aws-default-region` if not
-    /// using the fallback region.
-    ///
-    /// Prefer the environment variable over the command line flag in shared
-    /// environments.
-    #[structopt(long = "--aws-access-key-id", env = "AWS_ACCESS_KEY_ID")]
-    pub aws_access_key_id: Option<String>,
-
-    /// When using Amazon S3 as the object store, set this to the secret access
-    /// key that goes with the specified access key ID.
-    ///
-    /// Must also set `--object-store=s3`, `--bucket`, `--aws-access-key-id`.
-    /// Can also set `--aws-default-region` if not using the fallback region.
-    ///
-    /// Prefer the environment variable over the command line flag in shared
-    /// environments.
-    #[structopt(long = "--aws-secret-access-key", env = "AWS_SECRET_ACCESS_KEY")]
-    pub aws_secret_access_key: Option<String>,
-
-    /// When using Amazon S3 as the object store, set this to the region
-    /// that goes with the specified bucket if different from the fallback
-    /// value.
-    ///
-    /// Must also set `--object-store=s3`, `--bucket`, `--aws-access-key-id`,
-    /// and `--aws-secret-access-key`.
-    #[structopt(
-    long = "--aws-default-region",
-    env = "AWS_DEFAULT_REGION",
-    default_value = FALLBACK_AWS_REGION,
-    )]
-    pub aws_default_region: String,
-
-    /// When using Amazon S3 compatibility storage service, set this to the
-    /// endpoint.
-    ///
-    /// Must also set `--object-store=s3`, `--bucket`. Can also set `--aws-default-region`
-    /// if not using the fallback region.
-    ///
-    /// Prefer the environment variable over the command line flag in shared
-    /// environments.
-    #[structopt(long = "--aws-endpoint", env = "AWS_ENDPOINT")]
-    pub aws_endpoint: Option<String>,
-
-    /// When using Amazon S3 as an object store, set this to the session token. This is handy when using a federated
-    /// login / SSO and you fetch credentials via the UI.
-    ///
-    /// Is it assumed that the session is valid as long as the IOx server is running.
-    ///
-    /// Prefer the environment variable over the command line flag in shared
-    /// environments.
-    #[structopt(long = "--aws-session-token", env = "AWS_SESSION_TOKEN")]
-    pub aws_session_token: Option<String>,
-
-    /// When using Google Cloud Storage as the object store, set this to the
-    /// path to the JSON file that contains the Google credentials.
-    ///
-    /// Must also set `--object-store=google` and `--bucket`.
-    #[structopt(long = "--google-service-account", env = "GOOGLE_SERVICE_ACCOUNT")]
-    pub google_service_account: Option<String>,
-
-    /// When using Microsoft Azure as the object store, set this to the
-    /// name you see when going to All Services > Storage accounts > `[name]`.
-    ///
-    /// Must also set `--object-store=azure`, `--bucket`, and
-    /// `--azure-storage-access-key`.
-    #[structopt(long = "--azure-storage-account", env = "AZURE_STORAGE_ACCOUNT")]
-    pub azure_storage_account: Option<String>,
-
-    /// When using Microsoft Azure as the object store, set this to one of the
-    /// Key values in the Storage account's Settings > Access keys.
-    ///
-    /// Must also set `--object-store=azure`, `--bucket`, and
-    /// `--azure-storage-account`.
-    ///
-    /// Prefer the environment variable over the command line flag in shared
-    /// environments.
-    #[structopt(long = "--azure-storage-access-key", env = "AZURE_STORAGE_ACCESS_KEY")]
-    pub azure_storage_access_key: Option<String>,
 
     /// When IOx nodes need to talk to remote peers they consult an internal remote address
     /// mapping. This mapping is populated via API calls. If the mapping doesn't produce
@@ -300,18 +176,6 @@ fn parse_socket_addr(s: &str) -> std::io::Result<SocketAddr> {
     Ok(addrs
         .next()
         .expect("name resolution should return at least one address"))
-}
-
-arg_enum! {
-    #[derive(Debug, Copy, Clone, PartialEq)]
-    pub enum ObjectStore {
-        Memory,
-        MemoryThrottled,
-        File,
-        S3,
-        Google,
-        Azure,
-    }
 }
 
 #[cfg(test)]
